requirements.txt
# Core
streamlit==1.28.1
python-dotenv==1.0.0

# AI
google-generativeai==0.3.0

# Database
supabase==2.3.0

# Auth
PyJWT==2.8.1
cryptography==41.0.7

# Utils
requests==2.31.0
pandas==2.1.3